SAM debug: remove duplicate tests
awsSamDebugger.test.ts is a new file included in the parent (merge)
commit.  Some of its tests are the same as tests in
samDebugConfigProvider.test.ts.

TODO: The files can probably be combined later.
